[NO-JIRA] Adopt new api structure
