+++ helpers/common/apiStructure.js
@@ -1,4 +1,5 @@
 import appConfig from 'helpers/appConfig';
+import { apiStructureExtensions } from '../apiStructure_Extensions';
 
 const GET = 'GET';
 const POST = 'POST';
@@ -18,6 +19,10 @@ const apiStructure = {
     isSensitive: true,
   },
   verifyEmail: { method: POST, path: '/auth/verify-email', isSensitive: true },
+
+  // Emails
+  loadEmails: { method: GET, path: '/emails/load' },
+  resendEmail: { method: POST, path: '/emails/resend' },
 
   // Login
   loginWithMagicLink: { method: POST, path: '/magic-links/login', isSensitive: true },
@@ -44,30 +49,7 @@ const apiStructure = {
   signUp: { method: POST, path: '/users/sign-up', isSensitive: true },
   updateUser: { method: POST, path: '/users/update' },
 
-  // #region app-specific
-  // Blogs
-  createBlog: { method: POST, path: '/blogs/create' },
-  deleteBlog: { method: POST, path: '/blogs/delete' },
-  loadBlogs: { method: GET, path: '/blogs/load-all' },
-  loadBlog: { method: GET, path: '/blogs/load/:id' },
-  updateBlog: { method: POST, path: '/blogs/update' },
-
-  // Make payment
-  loadPayment: { method: POST, path: '/make-payment/load' },
-  createPaymentIntent: { method: POST, path: '/make-payment/create-intent' },
-  // TODO for admins to resend receipt
-  resendPaymentReceipt: { method: POST, path: '/make-payment/resend-receipt' },
-
-  // Payments
-  createPayment: { method: POST, path: '/payments/create' },
-  deletePayment: { method: POST, path: '/payments/delete' },
-  loadPayments: { method: GET, path: '/payments/load-all' },
-
-  // Support
-  createSupport: { method: POST, path: '/support/create' },
-  deleteSupport: { method: POST, path: '/support/delete' },
-  loadSupport: { method: GET, path: '/support/load' },
-  // #endregion app-specific
+  ...apiStructureExtensions,
 };
 
 Object.keys(apiStructure).forEach(key => {
