+++ src/containers/common/Authenticator/saga.js
@@ -1,7 +1,7 @@
 import { takeLatest, put, call } from 'redux-saga/effects';
 
 import request from 'utils/common/request';
-import apiStructure from 'helpers/apiStructure';
+import apiStructure from 'helpers/common/apiStructure';
 
 import { loadAuth } from './actions';
